+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [5.0.0] - 2026-05-18
 ### Added
 - TypeScript declarations.
 - Added separate ESM exports: default `imageBlobReduce` factory and named
@@ -140,7 +140,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First release.
 
 
-[Unreleased]: https://github.com/nodeca/image-blob-reduce/compare/4.1.0...HEAD
+[5.0.0]: https://github.com/nodeca/image-blob-reduce/compare/4.1.0...5.0.0
 [4.1.0]: https://github.com/nodeca/image-blob-reduce/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/nodeca/image-blob-reduce/compare/3.0.1...4.0.0
 [3.0.1]: https://github.com/nodeca/image-blob-reduce/compare/3.0.0...3.0.1